WIP
Former-commit-id: 4f010ab0343d3323c8e1ec1d83d52576870a45d6
Former-commit-id: f5ce2bda55f0b053bf2591d138c518975fcef190
Former-commit-id: 98b3b0943f3907f2eadeb9cbe2497a14041c5716

+++ Library/Core/Source/Exception.cpp
@@ -1,7 +1,4 @@
-#include <KAI/Core/Config/Base.h>
-#include <KAI/Core/Exception/ExceptionBase.h>
-#include <KAI/Core/StringStream.h>
-#include <KAI/Core/Type/Number.h>
+#include <KAI/Core/BuiltinTypes.h>
 
 KAI_BEGIN
 
@@ -29,7 +26,7 @@ namespace Exception
 
 	void UnknownObject::WriteExtendedInformation(StringStream &S) const
 	{
-		S << "handle=" << handle.GetValue();
+		S << "handle=" << (int)handle.GetValue();
 	}
 
 	void ObjectNotFound::WriteExtendedInformation(StringStream &S) const
